services/alert-sink/requirements.txt
<<<<<<< HEAD
psycopg2-binary
confluent-kafka
=======
confluent-kafka
psycopg2-binary
jsonschema
>>>>>>> f33ed383d8e88d290a27dd7885af588db7e1ce40
